DEVREL-1253 Update next and rewrite to app router
* update to new next version

* update to new next

* update api route to route handlers

* add dev:https script to package.json

* update homepage page to app router

* rewrite [slug] page into app router

* rewrite articles listing page

* rewrite articles [slug] to app router

* update products listing page to app router

* rewrite produicts detail page to app router

* rewrite solutions page to app router

* update callback and getPreviewApiKey to app router, remove pages router

* use new searchParam util in product listing

* use notFound instead of throwing when siteMenu can not be found

* remove unused funcitons, rewrite homepage route to use draft mode

* temporarily change to default eslint

* adjust readme

* remove unnecesray comment

* return null instead react fragment in smartlinkinitailzer
