--- .eslintrc.json
@@ -1,10 +1,0 @@
-{
-  "extends": [
-    "next/core-web-vitals",
-    "@kontent-ai",
-    "@kontent-ai/eslint-config/react"
-  ],
-  "rules": {
-    "@typescript-eslint/no-unused-vars": [2, { "ignoreRestSiblings": true }]
-  }
-}
