--- lib/utils/parseBoolean.ts
@@ -1,6 +1,0 @@
-export const parseBoolean = (str: string | string[] | undefined) => {
-    if (typeof str !== "string" || !["true", "false"].includes(str)) {
-      return null;
-    }
-    return str === "true";
-  }
