+++ package.json
@@ -3,14 +3,13 @@
   "version": "1.0.0",
   "scripts": {
     "generateModels": "tsc --project scripts/tsconfig.json && node scripts/build/generateModels.mjs",
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
+    "dev:https": "next dev --turbopack --experimental-https",
     "lint": "next lint",
     "lint:fix": "next lint --fix",
     "build": "next build",
     "export": "next export",
     "start": "next start",
-    "https:dev": "concurrently --kill-others -n dev,proxy \"npm run dev -- --port 3000\" \"local-ssl-proxy --source 3001 --target 3000\"",
-    "https:proxy": "local-ssl-proxy --source 3001 --target 3000",
     "test": "jest"
   },
   "dependencies": {
@@ -18,39 +17,44 @@
     "@kontent-ai/delivery-sdk": "^14.0.1",
     "@kontent-ai/rich-text-resolver": "^0.0.4",
     "@kontent-ai/smart-link": "^3.0.0",
-    "@portabletext/react": "^2.0.2",
-    "@vercel/analytics": "^1.0.0",
+    "@portabletext/react": "^3.2.0",
+    "@vercel/analytics": "^1.4.1",
     "auth0-js": "^9.22.1",
     "cookies-next": "^4.0.0",
     "flatted": "^3.2.9",
-    "next": "^13.5.3",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "next": "15.1.3",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
     "swiper": "^10.2.0"
   },
   "devDependencies": {
-    "@kontent-ai/eslint-config": "^0.1.3",
+    "@eslint/eslintrc": "^3.2.0",
     "@kontent-ai/management-sdk": "^5.0.0",
     "@kontent-ai/model-generator": "6.5.0",
     "@netlify/plugin-nextjs": "^4.40.1",
+    "@next/eslint-plugin-next": "^15.1.4",
     "@portabletext/types": "^2.0.6",
     "@tailwindcss/typography": "^0.5.9",
     "@types/auth0-js": "^9.21.0",
     "@types/jest": "^29.5.3",
     "@types/node": "^18.11.8",
-    "@types/react": "^18.0.24",
+    "@types/react": "19.0.2",
     "autoprefixer": "^10.4.14",
     "babel-jest": "^29.6.2",
     "concurrently": "^8.2.1",
     "dotenv": "^16.0.3",
-    "eslint": "^8.26.0",
-    "eslint-config-next": "^13.0.0",
+    "eslint": "^9.18.0",
+    "eslint-config-next": "15.1.4",
+    "eslint-plugin-react-hooks": "^5.1.0",
     "jest": "^29.6.2",
     "local-ssl-proxy": "^2.0.5",
     "postcss": "^8.4.22",
     "rimraf": "^5.0.0",
     "tailwindcss": "^3.3.1",
     "ts-jest": "^29.1.1",
-    "typescript": "5.0.4"
+    "typescript": "5.7.2"
+  },
+  "overrides": {
+    "@types/react": "19.0.2"
   }
 }
